# swagger.yml
openapi: 3.0.0
info:
  title: Smart Tool Log Parser API
  description: |
    Complete API documentation for the Smart Tool Log Parser system.
    
    This system processes manufacturing log files through multiple stages:
    1. **Ingest** - Receive and store files
    2. **Parse** - Extract structured data
    3. **Normalize** - AI-powered categorization with confidence scoring
    4. **Route** - Send to hot/cold paths based on urgency
    5. **Query** - Natural language search interface
    
    ## Services
    - **Pipeline API** (port 8080): synchronous `POST /process`, health; async job APIs are roadmap
    - **Query API** (port 8081): natural language search, schema
    - **WebSocket Alerts**: real-time P0/P1 (roadmap)
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/NAISCxSMU
servers:
  - url: http://localhost:8080
    description: Pipeline API (Production/Local)
  - url: http://localhost:8081
    description: Query API (Production/Local)
paths:
...
  /process:
    post:
      tags:
        - Pipeline - File Upload
      summary: Upload and process a log file (synchronous)
      description: |
        Runs ingest → parse → normalize → route → persist in a single request and returns a `ProcessResult`.
        HTTP status: 200 success, 206 partial_success (warnings in `errors`), 400 failed or empty file.
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            description: Optional format hint; omit for auto-detect
            enum: [json, xml, csv, log, txt]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Log file (JSON, XML, CSV, LOG, TXT)
              required:
                - file
      responses:
        "200":
          description: Processing completed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessResult"
        "206":
          description: Completed with warnings (partial_success)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessResult"
        "400":
          description: Failed processing or bad request (empty file, etc.)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProcessResult"
                  - type: object
                    properties:
                      detail:
                        type: string
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /preview:
    post:
      tags:
        - Pipeline - File Upload
      summary: Parse-only preview (no ingest / Kafka / AI)
      description: |
        Returns parsed records and errors from raw log text. Size- and rate-limited.
        Use for fast UI feedback before running the full `POST /process` pipeline.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: Raw log content (UTF-8)
                format:
                  type: string
                  nullable: true
                  description: Optional format hint (json, xml, csv, log, txt); omit or null for auto-detect
              required:
                - text
      responses:
        "200":
          description: Parse preview result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsePreviewOut"
        "413":
          description: Payload too large
        "429":
          description: Rate limit exceeded
  /status/{job_id}:
    get:
      tags:
        - Pipeline - Job Tracking
      summary: Job status (placeholder)
      description: |
        **Current implementation:** returns a placeholder payload. Persistent job tracking is roadmap.
        Use `job_id` from the synchronous `POST /process` response for correlation only.
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Placeholder status object
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                  note:
                    type: string
...
components:
  schemas:
    ProcessResult:
      type: object
      description: Synchronous pipeline outcome from POST /process
      properties:
        job_id:
          type: string
          format: uuid
        file_name:
          type: string
        file_format:
          type: string
          description: Detected or hinted format (e.g. json, csv, log)
        status:
          type: string
          enum: [success, partial_success, failed]
        events_processed:
          type: integer
        events_routed:
          type: object
          additionalProperties:
            type: integer
          description: Counts per Kafka topic (e.g. logs.p0)
        events_in_review:
          type: integer
        errors:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
        records_preview:
          type: array
          nullable: true
          description: Sample of normalized events (cap configured server-side)
          items:
            type: object
            additionalProperties: true
      required:
        - job_id
        - file_name
        - file_format
        - status
        - events_processed
        - events_routed
        - events_in_review
        - errors
        - timestamp
    ParsePreviewOut:
      type: object
      properties:
        detected_format:
          type: string
        record_count:
          type: integer
        records:
          type: array
          items:
            type: object
            additionalProperties: true
        parse_errors:
          type: array
          items: {}
        input_char_count:
          type: integer
        input_approx_tokens:
          type: number
          format: float
        tokenizer_version:
          type: string
...